View course

+++ src/app/admin/announcement/list/list.component.ts
@@ -42,18 +42,35 @@ export class ListComponent {
   }
 
   deleteAnnouncement(announcementId: any) {
-    this.announcementService.deleteAnnouncement(announcementId).subscribe((res: any) => {
-      Swal.fire({
-        title: 'Successful',
-        text: "Announcement deleted successfully",
-        icon: 'success',
-      });
 
-      this.activatedRoute.queryParams.subscribe(params => {
-        this.getAnnouncementList(params);
-      });
-      this.cdr.detectChanges();
+    
+
+    Swal.fire({
+      title: "Confirm Deletion",
+      text: "Are you sure you want to delete?",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#d33",
+      cancelButtonColor: "#3085d6",
+      confirmButtonText: "Delete",
+      cancelButtonText: "Cancel",
+    }).then((result) => {
+      if (result.isConfirmed){
+        this.announcementService.deleteAnnouncement(announcementId).subscribe((res: any) => {
+          Swal.fire({
+            title: 'Successful',
+            text: "Announcement deleted successfully",
+            icon: 'success',
+          });
+    
+          this.activatedRoute.queryParams.subscribe(params => {
+            this.getAnnouncementList(params);
+          });
+          this.cdr.detectChanges();
+        });
+      }
     });
+    
 
   }
