+++ src/app/admin/dashboard/main/main.component.ts
@@ -267,14 +267,29 @@ export class MainComponent implements OnInit {
         });
         return;
       }
-      this.classService.deleteClass(id).subscribe(() => {
-        Swal.fire({
-          title: 'Success',
-          text: 'Class deleted successfully.',
-          icon: 'success',
-        });
-        this.getClassList();
+
+      Swal.fire({
+        title: "Confirm Deletion",
+        text: "Are you sure you want to delete?",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#d33",
+        cancelButtonColor: "#3085d6",
+        confirmButtonText: "Delete",
+        cancelButtonText: "Cancel",
+      }).then((result) => {
+        if (result.isConfirmed){
+          this.classService.deleteClass(id).subscribe(() => {
+            Swal.fire({
+              title: 'Success',
+              text: 'Class deleted successfully.',
+              icon: 'success',
+            });
+            this.getClassList();
+          });
+        }
       });
+     
     });
   }
