+++ src/app/admin/program/program-list/program-list.component.ts
@@ -98,14 +98,29 @@ export class ProgramListComponent {
         });
         return;
       }
-      this.courseService.deleteProgram(id).subscribe(() => {
-        this.getProgramList();
-        Swal.fire({
-          title: 'Success',
-          text: 'Program deleted successfully.',
-          icon: 'success',
-        });
+
+      Swal.fire({
+        title: "Confirm Deletion",
+        text: "Are you sure you want to delete?",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#d33",
+        cancelButtonColor: "#3085d6",
+        confirmButtonText: "Delete",
+        cancelButtonText: "Cancel",
+      }).then((result) => {
+        if (result.isConfirmed){
+          this.courseService.deleteProgram(id).subscribe(() => {
+            this.getProgramList();
+            Swal.fire({
+              title: 'Success',
+              text: 'Program deleted successfully.',
+              icon: 'success',
+            });
+          });
+        }
       });
+      
     });
   }
   pageSizeChange($event: any) {
@@ -204,21 +219,36 @@ performSearch() {
   }
   removeSelectedRows() {
     const totalSelect = this.selection.selected.length;
-    this.selection.selected.forEach((item) => {
-      const index: number = this.programData.renderedData.findIndex(
-        (d: ProgramCourse) => d === item
-      );
-      // console.log(this.dataSource.renderedData.findIndex((d) => d === item));
-      // this.exampleDatabase?.dataChange.value.splice(index, 1);
-      this.refreshTable();
-      this.selection = new SelectionModel<ProgramCourse>(true, []);
-    });
+
     Swal.fire({
-      title: 'Success',
-      text: 'Record Deleted Successfully...!!!',
-      icon: 'success',
-      // confirmButtonColor: '#526D82',
+      title: "Confirm Deletion",
+      text: "Are you sure you want to delete?",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#d33",
+      cancelButtonColor: "#3085d6",
+      confirmButtonText: "Delete",
+      cancelButtonText: "Cancel",
+    }).then((result) => {
+      if (result.isConfirmed){
+        this.selection.selected.forEach((item) => {
+          const index: number = this.programData.renderedData.findIndex(
+            (d: ProgramCourse) => d === item
+          );
+          // console.log(this.dataSource.renderedData.findIndex((d) => d === item));
+          // this.exampleDatabase?.dataChange.value.splice(index, 1);
+          this.refreshTable();
+          this.selection = new SelectionModel<ProgramCourse>(true, []);
+        });
+        Swal.fire({
+          title: 'Success',
+          text: 'Record Deleted Successfully...!!!',
+          icon: 'success',
+          // confirmButtonColor: '#526D82',
+        });
+      }
     });
+   
     // this.showNotification(
     //   'snackbar-danger',
     //   totalSelect + ' Record Delete Successfully...!!!',
